requirements.txt
streamlit
streamlit_chat
sentence_transformers
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.3
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.1
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib